backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
requests==2.31.0
pydantic==2.7.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9